Downloader fixes

+++ GameLauncher/App/Classes/Downloader/ShowExtract.cs
@@ -1,5 +1,5 @@
 ﻿using System;
 
 namespace GameLauncher {
-	public delegate void ShowExtract(string filename, int currentCount, int allFilesCount);
+	public delegate void ShowExtract(string filename, long currentCount, long allFilesCount);
 }
